2단계 - 볼링 점수판(그리기) (#211)
* doc 요구사항 작성

* refactor step1 수정사항 반영

* feat 볼링 프레임별 게임 횟수 추가

* feat player 클래스 추가

* refactor Frame 클래스 수정 normarl frame 추가

* feat 볼링 점수 UI 구현

* feat 볼링 점수판 ui 구현

* fix 마지막 프레임 오류 수정

* feat view 클래스 추가

* fix 프레임 테스트 프린트 함수 제거

+++ src/main/java/qna/domain/Question.java
@@ -90,16 +90,19 @@ public class Question extends AbstractEntity {
     }
 
     public DeleteHistories delete(User user) throws CannotDeleteException {
-        DeleteHistories deleteHistories = new DeleteHistories();
-        deleteValidation(user);
+        validateOwner(user);
+
         this.deleted = true;
+
+        DeleteHistories deleteHistories = new DeleteHistories();
         deleteHistories.addDeleteHistory(new DeleteHistory(ContentType.QUESTION, getId(), getWriter()));
+
         Answers.of(getAnswers()).deleteALL(user, deleteHistories);
 
         return deleteHistories;
     }
 
-    private void deleteValidation(User user) throws CannotDeleteException {
+    private void validateOwner(User user) throws CannotDeleteException {
         if (!isOwner(user)) {
            throw new CannotDeleteException("질문을 삭제할 권한이 없습니다.");
         }
